% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_identical_rating.R
\name{maxIdenticalRating}
\alias{maxIdenticalRating}
\title{MaxIdenticalRating}
\usage{
maxIdenticalRating(
  x,
  varnames = NULL,
  verbose = TRUE,
  clean_data = TRUE,
  round_digits = 2
)
}
\arguments{
\item{x}{dataframe or tibble containing survey responses}

\item{varnames}{vector of variable names}

\item{verbose}{should the function output extra information, like how it
handles missing data or for appropriate measures, the battery-level
statistic.}

\item{clean_data}{should strings and factors be converted to numbers?
Defaults to TRUE. (These functions only work on numeric columns)}

\item{round_digits}{a numeric value specifying how many significant
digits you want}
}
\value{
a vector of the same length as the input dataframe that
indicates the proportion of answers which take the most common
value. Ranges from 0 to 1.
}
\description{
Calculates the proportion of answers taken up by the
most common response by the respondent to this set of questions.
Ignores NAs by default.
}
\examples{

survey <- data.frame("respondent" = c("A", "B"),
"item1" = c(5, 5),
"item2" = c(5, 5),
"item3" = c(5, 5),
"item4" = c(5, 1))

maxIdenticalRating(survey, varnames = c("item1", "item4"))
}
