% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRP.R
\name{meanRootPairs}
\alias{meanRootPairs}
\title{Mean Root Pairs}
\usage{
meanRootPairs(
  x,
  varnames = NULL,
  verbose = TRUE,
  clean_data = TRUE,
  round_digits = 2
)
}
\arguments{
\item{x}{dataframe or tibble containing survey responses}

\item{varnames}{vector of variable names}

\item{verbose}{should the function output extra information, like how it
handles missing data or for appropriate measures, the battery-level
statistic.}

\item{clean_data}{should strings and factors be converted to numbers?
Defaults to TRUE. (These functions only work on numeric columns)}

\item{round_digits}{a numeric value specifying how many significant
digits you want}
}
\value{
a vector containing the measure for each row of the dataset calculated
across the variables specified in \code{varnames}
}
\description{
Calculate the mean Root Pairs measure for each respondent
over a set of supplied variables
}
\examples{

survey <- data.frame("respondent" = c("A", "B"),
"item1" = c(5, 5),
"item2" = c(5, 5),
"item3" = c(5, 5),
"item4" = c(5, 1))


meanRootPairs(survey, varnames = c("item1", "item4"))
}
