% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_nondiff.R
\name{nondiff}
\alias{nondiff}
\title{Title}
\usage{
nondiff(
  x,
  varnames = NULL,
  verbose = TRUE,
  clean_data = TRUE,
  round_digits = 2
)
}
\arguments{
\item{x}{dataframe or tibble containing survey responses}

\item{varnames}{vector of variable names}

\item{verbose}{should the function output extra information, like how it
handles missing data or for appropriate measures, the battery-level
statistic.}

\item{clean_data}{should strings and factors be converted to numbers?
Defaults to TRUE. (These functions only work on numeric columns)}

\item{round_digits}{a numeric value specifying how many significant
digits you want}
}
\value{
vector that indicates whether
a given respondent differs in their answers or not. TRUE indicates no differences
between answers to this question
}
\description{
Title
}
\examples{

survey <- data.frame("respondent" = c("A", "B"),
"item1" = c(5, 5),
"item2" = c(5, 5),
"item3" = c(5, 5),
"item4" = c(5, 1))

nondiff(survey, varnames = c("item1", "item4"))
}
