% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_sd.R
\name{question_sd}
\alias{question_sd}
\title{Standard Deviation}
\usage{
question_sd(
  x,
  varnames = NULL,
  verbose = TRUE,
  clean_data = TRUE,
  round_digits = 2
)
}
\arguments{
\item{x}{dataframe or tibble containing survey responses}

\item{varnames}{vector of variable names}

\item{verbose}{should the function output extra information, like how it
handles missing data or for appropriate measures, the battery-level
statistic.}

\item{clean_data}{should strings and factors be converted to numbers?
Defaults to TRUE. (These functions only work on numeric columns)}

\item{round_digits}{a numeric value specifying how many significant
digits you want}
}
\value{
vector of standard deviation for this set of questions
}
\description{
Standard Deviation
}
\details{
NAs are ignored in the calculation
}
\examples{

survey <- data.frame("respondent" = c("A", "B"),
"item1" = c(5, 5),
"item2" = c(5, 5),
"item3" = c(5, 5),
"item4" = c(5, 1))

question_sd(survey, varnames = c("item1", "item3", "item4"))
}
