% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/straightlining-package.R
\docType{package}
\name{straightliner-package}
\alias{straightliner}
\alias{straightliner-package}
\title{straightliner: Check Survey Responses Using a Variety of Straightlining Measures}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package implements several quantiative measures of straigtlining as descriped in Kim et al. (2019). It offers a user-friendly way to assess whether individual are straightlining, or compare sets of questions against one another to see if people straightline more in some batteries than in others.
}
\author{
\strong{Maintainer}: Max weylandt \email{mweylandt@gmail.com} (\href{https://orcid.org/0000-0001-8590-4145}{ORCID})

}
\keyword{internal}
