% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/straightlining_qset.R
\name{straightlining_qset}
\alias{straightlining_qset}
\title{straightlining_qset}
\usage{
straightlining_qset(
  x,
  variable_list = NULL,
  measures = c("mrp", "mir", "qsd", "spv", "nondiff"),
  round_digits = 2,
  verbose = FALSE
)
}
\arguments{
\item{x}{a data frame or tibble containing the responses you wish to analyze}

\item{variable_list}{a named list. Each list item should be a vector containing
the column names or indices of the questions you want a report on.}

\item{measures}{a vector of the measures you want a report on. See also \link{straightlining}}

\item{round_digits}{a numeric value specifying how many significant digits you want}

\item{verbose}{should the function output extra information, like how it
handles missing data or for appropriate measures, the battery-level
statistic.}
}
\value{
a data frame containing the measures for every set of questions you
requested
}
\description{
Get a report aggregate straightlining measures for (several) batteries of
questions
}
\details{
Usually, you want to see whether individuals are straightlining to see
whether they are of high-enough quality to include in the analysis. But some
questions are also more stimulating than others, and it might be useful to
see whether some questions encourage respondents to engage in straightlining.
(this should go in a vignette)

Simple nondifferentiation ('nondiff') and Mean Root Pairs ('mrp') are by their
nature designed to report on a set of questions rather than on individual
respondents. I also give you the option to return averages for standard
deviation ('sd'), scale point variation ('spv'), and maximum Identical Rating ('mir')
Others have certainly averaged Scale Point Variation in this way before
(Kim et al. 2019, p.19 CHECK THIS), but for all three of these I make no guarantee that the
mean is an appropriate way of aggregating results.
}
\examples{
survey <- data.frame("respondent" = c("A", "B"),
"item1" = c(5, 5),
"item2" = c(5, 5),
"item3" = c(5, 5),
"item4" = c(5, 1))

batteries <- list(first_two = c("item1", "item2"), second_2 = c(3,4))
straightlining_qset(survey, variable_list = batteries)
}
\references{
Kim, Y., Dykema, J., Stevenson, J., Black, P., & Moberg, D. P. (2019).
Straightlining: Overview of measurement, comparison of indicators, and
effects in mail–web mixed-mode surveys. Social Science Computer Review,
37(2), 214-233.
}
